+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+#### Version 0.0.5+5 (2021-05-11) 
+* Add coingecko api (Get Uniris Coin infos (name, price, market, ... including exchange tickers))
+* Add absinthe socket
+
 #### Version 0.0.4+4 (2021-05-04) 
 * Add api requests (fetchBalance, getTransactionContent, getTransactions)
 * Add coingecko api